add openDeviceContactPicker functionality
